# workspaces/backend/openapi/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage notebooks in a Kubernetes cluster.\nFor more information, visit https://www.kubeflow.org/docs/components/notebooks/",
        "title": "Kubeflow Notebooks API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/workspaces": {
            "get": {
                "description": "Returns a list of all workspaces in the cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspaces"
                ],
                "summary": "List all workspaces",
                "operationId": "listAllWorkspaces",
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns a list of all workspaces.",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to list workspaces.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/workspaces/{namespace}": {
            "get": {
                "description": "Returns a list of workspaces in a specific namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspaces"
                ],
                "summary": "List workspaces by namespace",
                "operationId": "listWorkspacesByNamespace",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "kubeflow-user-example-com",
                        "description": "Namespace to filter workspaces",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation. Returns a list of workspaces in the specified namespace.",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceListEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to list workspaces.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new workspace in the specified namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspaces"
                ],
                "summary": "Create workspace",
                "operationId": "createWorkspace",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "kubeflow-user-example-com",
                        "description": "Namespace for the workspace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace creation configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceCreateEnvelope"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Workspace created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceCreateEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to create workspace.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict. Workspace with the same name already exists.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large. The request body is too large.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type. Content-Type header is not correct.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/workspaces/{namespace}/{name}": {
            "put": {
                "description": "Updates an existing workspace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workspaces"
                ],
                "summary": "Update workspace",
                "operationId": "updateWorkspace",
                "parameters": [
                    {
                        "type": "string",
                        "x-example": "kubeflow-user-example-com",
                        "description": "Namespace of the workspace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "x-example": "my-workspace",
                        "description": "Name of the workspace",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace update configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workspace updated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.WorkspaceEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden. User does not have permission to update workspace.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict. Current workspace revision is newer than provided.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large. The request body is too large.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type. Content-Type header is not correct.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity. Validation error.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal server error. An unexpected error occurred on the server.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorEnvelope": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/api.HTTPError"
                }
            }
        },
        "api.HTTPError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "cause": {
                    "description": "Cause contains detailed information about the cause of the error.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ErrorCause"
                        }
                    ]
                },
                "code": {
                    "description": "Code is a string representation of the HTTP status code.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a human-readable description of the error.",
                    "type": "string"
                }
            }
        },
...
        "api.WorkspaceCreateEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/workspaces.WorkspaceCreate"
                }
            }
        },
        "api.WorkspaceEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/workspaces.WorkspaceUpdate"
                }
            }
        },
...
        "api.WorkspaceListEnvelope": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspaces.WorkspaceListItem"
                    }
                }
            }
        },
...
        "workspaces.Service": {
            "type": "object",
            "properties": {
                "httpService": {
                    "$ref": "#/definitions/workspaces.HttpService"
                }
            }
        },
        "workspaces.WorkspaceCreate": {
            "type": "object",
            "required": [
                "kind",
                "name",
                "paused",
                "podTemplate"
            ],
            "properties": {
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paused": {
                    "type": "boolean"
                },
                "podTemplate": {
                    "$ref": "#/definitions/workspaces.PodTemplateMutate"
                }
            }
        },
...
        "workspaces.WorkspaceListItem": {
            "type": "object",
            "required": [
                "activity",
                "audit",
                "name",
                "namespace",
                "paused",
                "pausedTime",
                "pendingRestart",
                "podTemplate",
                "services",
                "state",
                "stateMessage",
                "workspaceKind"
            ],
            "properties": {
                "activity": {
                    "$ref": "#/definitions/workspaces.Activity"
                },
                "audit": {
                    "$ref": "#/definitions/common.Audit"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "paused": {
                    "type": "boolean"
                },
                "pausedTime": {
                    "type": "integer"
                },
                "pendingRestart": {
                    "type": "boolean"
                },
                "podTemplate": {
                    "$ref": "#/definitions/workspaces.PodTemplate"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/workspaces.Service"
                    }
                },
                "state": {
                    "$ref": "#/definitions/v1beta1.WorkspaceState"
                },
                "stateMessage": {
                    "type": "string"
                },
                "workspaceKind": {
                    "$ref": "#/definitions/workspaces.WorkspaceKindInfo"
                }
            }
        },
...
    }
}